Replace OS X libgtest.a with what works on my environment
The current binary in the tree doesn't work on my environment.
Also, it isn't clear on which environment it works at all.

This binary is built by me from googletest release-1.8.0 source
with the following environment. It's OS X Yosemite 10.10.5.
It's better than the current binary because it's known to work
on at least one environment.

    giraffe% uname -a
    Darwin giraffe.local 14.5.0 Darwin Kernel Version 14.5.0: Sun Jun  4 21:40:08 PDT 2017; root:xnu-2782.70.3~1/RELEASE_X86_64 x86_64
    giraffe% clang -v
    Apple LLVM version 7.0.2 (clang-700.1.81)
    Target: x86_64-apple-darwin14.5.0
    Thread model: posix
    giraffe%

Of course, this is not a real fix.  We should not have added binaries
like this in the first place.  As it seems that this library has ABI
issues, we probably need to build it from the source during our build.

Ref: MI-2683
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: Ib81c4f4c6b25565ec9345d93aff9a846590cd60e