Added build support on ARM v8
This patch only enables package generation for ARMv8 architecture.

To compile ARMv8 packages midonet build must be run on armv8 linux, otherwise
x86_64 compatible packages are produced. There is no ARMv8 compatible
protobuf compilator in Maven, so it should be installed as a separte package.

Seems that it is possible to enable "cross compilation" somehow.

Unit tests support should be added in a separate patch.

Ref: MI-2706
Change-Id: Id2e33072b92236065dfef1309667fc5d9098a43c