[WIP] Flow state hash for ICMP in ICMP error
This fixes an issue where the MidoNet Agent would compute
the inverse connection hash (to determine the correct simulation
thread for a reverse packet) only for ICMP error packets due
to a forward transport packet (TCP or UDP).

However, ICMP error responses can also happen for ICMP forward
packets. This case was not previously handled.

This change addresses this, by setting the correct ICMP
identifier in the computed flow match when an ICMP error
reverse packet is due to an ICMP forward packet.

TODO: Unit tests

Ref: MI-2680

Change-Id: I43b458e1c415ca5198d7cde724bca606ad847252
Signed-off-by: Alex Bikfalvi <alex@bikfalvi.com>