Adding per host port binding to allow live-migration
During live migration in Openstack, at the pre-live migration phase,
the destination host binding info needs to be stored without
actually performing the bound on the destination.

Once Neutron has notified about the change of host for the VM port,
the binding information for the destination host is used.

This patch tries to fulfill these requirements by:

 * Creating a HostPortBinding entry in Zookeeper that will store the
   binding information (interface name) for a given host and port.
   The ID of the entry is the XOR of the IDs of host and port.
 * Modifying PortTranslator so that in case that the binding info is
   available on the host update for the port, the interface is also
   updated.
 * Modifying the HostMapper so that it completes the bindings from the
   data stored in the HostPortBinding entries, if available, and
   listen to changes in them.

Ref: MNA-1066

Change-Id: Ifa4cc31c980da6479ddb3e0083ab2222f7e73665
Signed-off-by: Miguel Herranz <miguel@midokura.com>