Defining HostPortBinding object
A new protocol buffer message is defined to declare the binding
of a specific host and port pair to an interface in that host.

The entries in Zookeeper are mapped to a new virtual device of
the same name.

The ID of the message is a mix between the host UUID and port
UUID. A new function has been added to UUIDUtils to handle it.

Ref: MNA-1066

Change-Id: I580e4900e98f9b49389f4e54444fd8cd8ed1fd11
Signed-off-by: Miguel Herranz <miguel@midokura.com>