Update patch set 3
Patch Set 3:

Thanks for the detailed explanation. If this change passes the unit tests, we can merge and start unblocking the queue until you provide the proper fix.  If you manage to confirm that the error occurs from a specific test that was introduced in 5.6 before that, we could just disable that test and abandon this one.

Patch-set: 3