TopologyCacheTest must close backend service
This fixes an issue in the TopologyCacheTest, where the test
did not close a local instance of the MidonetBackendService
after the test completion, whereas the test-local ZooKeeper
was being shutdown.

Closing the MidonetBackendService after a unit test is a
requirement, since the default behavior of the service when
losing the ZooKeeper connection is to exit the application
by calling System#exit(int) after the expiration of a
configurable grace time.

This, during CI, may lead to a premature exit of the application
while executing the unit tests, if the grace time expires before
the tests being completed.

Ref: MI-2710

Change-Id: I13ac09bb899d28354f2a078f4cc5b5d90f5a1599
Signed-off-by: Alex Bikfalvi <alex@bikfalvi.com>