Create MDTS '~/.midonetrc' on CI start script
Some scripts in Jenkins (full_mdts_* ones) are failing because the
TAP tests 01-setup_openstack.bats and 10-connectivity.bats cannot
connect to Midonet API since the configuration is missing.

That configuration can be provided by environment variables, but
since having a ~/.midonetrc for MDTS could be handy even for manual
tests this is created on CI start script so it can be used by the
TAP tests later.

Ref: MINOR_CHANGE

Change-Id: Ia11d069adf75ff70ccd4926277b35211b03ee504
Signed-off-by: Miguel Herranz <miguel@midokura.com>