Mark libreswan and vpp as recommended
This eliminates package conflicts when installing the Midolman package
in CPU architectures other than x64 (e.g. arm64), since libreswan and
vpp are not readily available in standard repositories.

After all, these packages are only needed if certain features of Midonet
are used.

We might need to do the same for the RHEL & SuSE packages when needed.

Change-Id: I5fde302e64fb81900e89add0aa71c993b62f61a1