edit MidoNet license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit modifies a few of MidoNet's docs so that Licensee is able
to recognize the repository's license type. It updates LICENSE so that
it contains the full text of the Apache license. It also updates the
README with a new "Licensing" section.

Collectively, these changes allow Licensee to successfully identify the
license type of MidoNet as Apache 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>