edit MidoNet license info for GitHub and Licensee
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit modifies a few of MidoNet's docs so that Licensee is able
to recognize the repository's license type. It updates LICENSE so that
it contains the full text of the Apache license. It also updates the
README with a new "Licensing" section.

Collectively, these changes allow Licensee to successfully identify the
license type of MidoNet as Apache 2.0.

Ref: MINOR_CHANGE
Change-Id: I7082146ae107e3d4a7db31a95347e600b4fb86ef
Signed-off-by: Andrea Kao <eirinikos@gmail.com>