Update patch set 6
Patch Set 6:

(4 comments)

It looks ok to me, but I don't know well this part of the code, so I am not sure why this change is this needed or what problem resolves (it seems that dpPortNumToTriad is only used to finally remove the port in the same way it is done by ensurePortPid already). See the comments in the code, maybe you can help me to understand it better.

Patch-set: 6
Reviewer: Miguel Herranz <1006717@d5d70762-12d0-45a1-890d-524b12d3f735>